ED/requirements.txt
keras==2.13.1
matplotlib==3.7.2
numpy==1.24.3
Pillow==10.0.0
python_Levenshtein==0.21.1
scipy==1.11.2
tensorflow==2.13.0
